% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{convertJSON}
\alias{convertJSON}
\title{convert nested json input to unnested format}
\usage{
convertJSON(input)
}
\arguments{
\item{input}{Input from opencpu}
}
\description{
convert nested json input to unnested format
}

