% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{unnest}
\alias{unnest}
\title{Find the candidates of bandwidths for locLinear2d}
\usage{
unnest(DT, unnestCols = NULL)
}
\arguments{
\item{DT}{A data.table containing list or vector in the cell.
The cells in each row must have the same number of elements.}

\item{unnestCols}{The column names to unnest.}
}
\value{
A unnested data.table.
}
\description{
Find the candidates of bandwidths for locLinear2d
}
\examples{
require(data.table)
DT <- unnest(data.table(V1 = list(c(1,3,5), c(1,7)), V2 = list(c(2,5,3), c(4,6)), V3 = 1:2))
}

